# 8322767: TLS full handshake is slow with PKCS12KeyStore and X509KeyManagerImpl