@johnshajiang 2024-03-19
@haimaychao 
Just a suggestion.
Could you please re-format those micro bench results to make them more readable?
If you still have the original JMH outputs, it's easy to to that with Markdown.