@XueleiFan 2024-03-26
> While attempting to handle when to refresh a cached entry to accommodate keystore changes, we would like to know if you agree that this improvement is worth the risk. We would also like to know if you have a preference for other options:

I'm not sure if it is the direction to improve performance by introducing new bugs.

If I read the benchmark result correction, for some cases, the performance was hurt by the update, some others was improved.  For the performance improved cases, it may be just in the cost of the key manager initialization performance.  The cost may be not desired as it hurts the boot-up performance to build a TLS context.

Alternatively, there might/could be an event get triggered when the key store get updated. It may worth to explore how to trigger the event, listen to the even and handle the event accordingly in the key manager and key store.