@XueleiFan 2024-03-27
From the description, it looks like you are concerning about that new bugs could be introduced and if the performance improvement worth the risk.  Before looking into the performance numbers, I think it would be nice to make sure no bugs was introduced. Otherwise, the performance improvement makes no sense.

> Without changes:
> SSLHandshake.doHandshake PKIX true TLS thrpt 15 937.166 ? 23.937 ops/s
>
> With changes:
> SSLHandshake.doHandshake PKIX true TLS thrpt 15 910.733 ? 87.172 ops/s

Here is a performance regression, from 937 to 910 ops/s.

> I created a JMH benchmark to measure the time needed to build a TLS context using different SSL protocols ("TLSv1.2" and "TLS") and different KeyManager types ("SunX509" and "PKIX”). Here is the benchmark result:
> 
> Benchmark (keyMgr) (tlsVersion) Mode Cnt Score Error Units SSLHandshake.doHandshake SunX509 TLSv1.2 thrpt 15 112.920 ? 10.518 ops/s SSLHandshake.doHandshake SunX509 TLS thrpt 15 122.255 ? 3.128 ops/s SSLHandshake.doHandshake PKIX TLSv1.2 thrpt 15 115.956 ? 5.490 ops/s SSLHandshake.doHandshake PKIX TLS thrpt 15 118.512 ? 3.574 ops/s
> 
> The performance between the SunX509 and PKIX KeyManagers for building SSL contexts seems fairly comparable. The changes made to X509KeyManagerImpl.java to cache keystore entries in the credentialsMap at initialization time for the PKIX KeyManager do not impact performance compared to SunX509.
> 

I think the comparing could between apple to apple.  PKIX and SunX509 are different key managers and could have different performance numbers.  Did you have the number for PKIX before and after the update for TLS boot-up?